Fix corresponding with Telegram Bot API

+++ src/jsMain/kotlin/com/kirillNay/telegram/miniapp/webApp/data/WebAppChat.kt
@@ -11,7 +11,7 @@ external class WebAppChat {
      * Unique identifier for this chat. This number may have more than 32 significant bits and some programming languages may have difficulty/silent defects in interpreting it. But it has at most 52 significant bits, so a signed 64-bit integer or double-precision float type are safe for storing this identifier.
      */
     @JsName("id")
-    val id: Int
+    val id: Long
 
     /**
      * Type of chat, can be either “group”, “supergroup” or “channel”
